Update validate BLL

+++ BLL/TableBLL.cs
@@ -2,9 +2,11 @@
 using DTO;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 
 namespace BLL
 {
@@ -28,20 +30,26 @@ namespace BLL
 
         public bool CheckTableExistsBill(int idTable)
         {
+            if ( string.IsNullOrEmpty(idTable.ToString()) || string.IsNullOrWhiteSpace(idTable.ToString()))
+            {
+                return false;
+            }
             Response res = (Response)TableDAL.Instance.CheckTableExistsBill(idTable);
             if (res.error == false) return true;
             else return false;
         }
-
         public bool InsertTable()
         {
             Response res = (Response)TableDAL.Instance.InsertTable();
             if (res.error == false) return true;
             else return false;
         }
-
         public bool DeleteTable(int idTable)
         {
+            if (string.IsNullOrEmpty(idTable.ToString()) || string.IsNullOrWhiteSpace(idTable.ToString()))
+            {
+                return false;
+            }
             Response res = (Response)TableDAL.Instance.DeleteTable(idTable);
             if (res.error == false) return true;
             else return false;
